refactor(stats): split components (StudentTransitionTab, ChangeRateText)
